requirements.txt
langgraph>=0.2.55
langchain-community>=0.3.9
langchain-core>=0.1.0
tavily-python>=0.3.4
openrouter>=0.3.0
python-dotenv==1.0.1
rich==13.7.0
